Move server files
