+++ src/server.js
@@ -3,8 +3,8 @@ const express = require('express');
 const compression = require('compression');
 const path = require('path');
 
-const { applyMiddlewares } = require('./src_server/middleware');
-const { applyRouters } = require('./src_server/router');
+const { applyMiddlewares } = require('./middleware');
+const { applyRouters } = require('./router');
 
 const app = express();
 const router = express.Router();
@@ -22,7 +22,7 @@ app.use(express.static('dist'));
 
 // redirect all requests to index.html
 app.use((req, res) => {
-  res.sendFile(path.resolve(__dirname, './dist/index.html'));
+  res.sendFile(path.resolve(__dirname, '../dist/index.html'));
 });
 
 // starting listening
